+++ app/Observers/AccountObserver.php
@@ -8,12 +8,12 @@ class AccountObserver
 {
     public function creating(Account $account): void
     {
-        $account->created_by = auth()->id();
+        $account->created_by = (int) auth()->id();
         $account->status = 'active';
     }
 
     public function updating(Account $account): void
     {
-        $account->updated_by = auth()->id();
+        $account->updated_by = (int) auth()->id();
     }
 }
