+++ app/Observers/AttributeObserver.php
@@ -9,12 +9,12 @@ class AttributeObserver
     public function creating(Attribute $attribute): void
     {
         $attribute->slug = str($attribute->name)->slug()->toString();
-        $attribute->created_by = auth()->id();
+        $attribute->created_by = (int) auth()->id();
     }
 
     public function updating(Attribute $attribute): void
     {
         $attribute->slug = str($attribute->name)->slug()->toString();
-        $attribute->updated_by = auth()->id();
+        $attribute->updated_by = (int) auth()->id();
     }
 }
