+++ app/Observers/DepositCategoryObserver.php
@@ -9,12 +9,12 @@ class DepositCategoryObserver
     public function creating(DepositCategory $depositCategory): void
     {
         $depositCategory->slug = str($depositCategory->name)->slug()->toString();
-        $depositCategory->created_by = auth()->id();
+        $depositCategory->created_by = (int) auth()->id();
     }
 
     public function updating(DepositCategory $depositCategory): void
     {
         $depositCategory->slug = str($depositCategory->name)->slug()->toString();
-        $depositCategory->updated_by = auth()->id();
+        $depositCategory->updated_by = (int) auth()->id();
     }
 }
