+++ app/Observers/SupplierObserver.php
@@ -8,11 +8,11 @@ class SupplierObserver
 {
     public function creating(Supplier $supplier): void
     {
-        $supplier->created_by = auth()->id();
+        $supplier->created_by = (int) auth()->id();
     }
 
     public function updating(Supplier $supplier): void
     {
-        $supplier->updated_by = auth()->id();
+        $supplier->updated_by = (int) auth()->id();
     }
 }
