+++ app/Observers/WarehouseObserver.php
@@ -9,12 +9,12 @@ class WarehouseObserver
     public function creating(Warehouse $warehouse): void
     {
         $warehouse->slug = str($warehouse->name)->slug()->toString();
-        $warehouse->created_by = auth()->id();
+        $warehouse->created_by = (int) auth()->id();
     }
 
     public function updating(Warehouse $warehouse): void
     {
         $warehouse->slug = str($warehouse->name)->slug()->toString();
-        $warehouse->updated_by = auth()->id();
+        $warehouse->updated_by = (int) auth()->id();
     }
 }
